Process L1 blocks in a different task. (#1163)
* Process L1 blocks in a different task.

This introduces handler for L1 blocks which is spawned as a separate task

* Add DA block handler

* Happy clippy

* Move all L1 processing into L1BlockHandler

* Remove uninformative log

* Keep logic for retrying L1 blocks if L2 range is missing

* Happy Clippy: 2

* Process L2 blocks in order

* add some logs

* Process blocks immediately and queue if failed

* Import info

* Fix logs

* Process all L2 blocks in pending

* more logs

---------
