+++ crates/sovereign-sdk/rollup-interface/src/state_machine/zk/mod.rs
@@ -65,7 +65,7 @@ pub trait ZkvmHost: Zkvm + Clone {
 /// Must support recursive proofs.
 pub trait Zkvm: Send + Sync {
     /// A commitment to the zkVM program which is being proven
-    type CodeCommitment: Clone + Debug + Serialize + DeserializeOwned;
+    type CodeCommitment: Clone + Debug + Serialize + DeserializeOwned + Send + Sync + 'static;
 
     /// The error type which is returned when a proof fails to verify
     type Error: Debug;
